split benchmarks and add integration benchmark
